refactor(ds): Move related components to product-components
