--- packages/components/src/components/TrezorLogo/trezorLogos.ts
@@ -1,8 +1,0 @@
-export const LOGOS: { [key: string]: any } = {
-    HORIZONTAL: require('../../images/trezor_logo_horizontal.svg'),
-    VERTICAL: require('../../images/trezor_logo_vertical.svg'),
-    SYMBOL: require('../../images/trezor_logo_symbol.svg'),
-    SUITE: require('../../images/trezor_logo_suite.svg'),
-    SUITE_SQUARE: require('../../images/trezor_logo_suite_square.svg'),
-    SUITE_COMPACT: require('../../images/trezor_logo_suite_compact.svg'),
-};
