+++ packages/product-components/src/components/PasswordStrengthIndicator/PasswordStrengthIndicator.stories.tsx
