+++ packages/product-components/src/components/PasswordStrengthIndicator/PasswordStrengthIndicator.tsx
