+++ packages/product-components/src/components/RotateDeviceImage/RotateDeviceImage.stories.tsx
