+++ packages/product-components/src/components/TrezorLogo/TrezorLogo.tsx
