+++ packages/product-components/src/components/TrezorLogo/trezorLogos.ts
@@ -1,0 +1,8 @@
+export const LOGOS: { [key: string]: any } = {
+    HORIZONTAL: require('../../../../components/src/images/trezor_logo_horizontal.svg'),
+    VERTICAL: require('../../../../components/src/images/trezor_logo_vertical.svg'),
+    SYMBOL: require('../../../../components/src/images/trezor_logo_symbol.svg'),
+    SUITE: require('../../../../components/src/images/trezor_logo_suite.svg'),
+    SUITE_SQUARE: require('../../../../components/src/images/trezor_logo_suite_square.svg'),
+    SUITE_COMPACT: require('../../../../components/src/images/trezor_logo_suite_compact.svg'),
+};
