+++ packages/product-components/src/index.ts
@@ -1,3 +1,6 @@
 export { PassphraseTypeCard } from './components/PassphraseTypeCard/PassphraseTypeCard';
 export { ConfirmOnDevice } from './components/ConfirmOnDevice/ConfirmOnDevice';
 export { mapTrezorModelToIcon } from './utils/mapTrezorModelToIcon';
+export { RotateDeviceImage } from './components/RotateDeviceImage/RotateDeviceImage';
+export { TrezorLogo } from './components/TrezorLogo/TrezorLogo';
+export { PasswordStrengthIndicator } from './components/PasswordStrengthIndicator/PasswordStrengthIndicator';
