+++ packages/suite/src/components/suite/layouts/SuiteLayout/DeviceSelector/DeviceStatus.tsx
@@ -5,7 +5,7 @@ import { DeviceDetail } from 'src/views/suite/SwitchDevice/DeviceItem/DeviceDeta
 import { MouseEventHandler } from 'react';
 import { TrezorDevice } from 'src/types/suite';
 import { spacingsPx } from '@trezor/theme';
-import { RotateDeviceImage } from '@trezor/components';
+import { RotateDeviceImage } from '@trezor/product-components';
 import { DeviceStatusText } from 'src/views/suite/SwitchDevice/DeviceItem/DeviceStatusText';
 import { selectDeviceLabelOrName } from '@suite-common/wallet-core';
 import { useSelector } from 'src/hooks/suite';
